fix(db): Replace erb content with nunjucks content
For each files I replace the erb syntax with the nunjucks syntax. To achieve this, I used sed like
this  and after I
renamed them like they should be. connectbot, roxterm and shell did not change because their
content are wrapped in function. You can see this with a grep .